requirements.txt
# external requirements
bentoml
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
kaggle
neptune-client
neptune-notebooks
matplotlib
scikit-learn
seaborn
swifter
lightgbm
scikit-plots
